requirements.txt
alembic==1.11.1
blinker==1.6.2
Bootstrap-Flask==2.2.0
click==8.1.3
colorama==0.4.6
dominate==2.7.0
Flask==2.3.2
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
visitor==0.1.3
Werkzeug==2.3.4
WTForms==3.0.1
